Split leveldb and memory blockchain storage
